객체지향 구조 변경
